+++ config/plugin-commands-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pnpm/plugin-commands-config
 
+## 2.0.30
+
+### Patch Changes
+
+- @pnpm/cli-utils@2.1.4
+
 ## 2.0.29
 
 ### Patch Changes