+++ packages/plugin-commands-doctor/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pnpm/plugin-commands-doctor
 
+## 2.0.29
+
+### Patch Changes
+
+- @pnpm/cli-utils@2.1.4
+
 ## 2.0.28
 
 ### Patch Changes