+++ packages/plugin-commands-setup/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pnpm/plugin-commands-setup
 
+## 4.0.29
+
+### Patch Changes
+
+- @pnpm/cli-utils@2.1.4
+
 ## 4.0.28
 
 ### Patch Changes