+++ patching/plugin-commands-patching/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @pnpm/plugin-commands-patching
 
+## 5.0.12
+
+### Patch Changes
+
+- @pnpm/plugin-commands-installation@14.0.12
+
+## 5.0.11
+
+### Patch Changes
+
+- Updated dependencies [6558d1865]
+  - @pnpm/plugin-commands-installation@14.0.11
+  - @pnpm/cli-utils@2.1.4
+  - @pnpm/store-connection-manager@7.0.18
+
 ## 5.0.10
 
 ### Patch Changes