+++ releasing/plugin-commands-deploy/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @pnpm/plugin-commands-deploy
 
+## 4.0.12
+
+### Patch Changes
+
+- @pnpm/plugin-commands-installation@14.0.12
+
+## 4.0.11
+
+### Patch Changes
+
+- Updated dependencies [6558d1865]
+  - @pnpm/plugin-commands-installation@14.0.11
+  - @pnpm/cli-utils@2.1.4
+
 ## 4.0.10
 
 ### Patch Changes