+++ reviewing/plugin-commands-licenses/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pnpm/plugin-commands-licenses
 
+## 3.0.7
+
+### Patch Changes
+
+- @pnpm/cli-utils@2.1.4
+
 ## 3.0.6
 
 ### Patch Changes