+++ store/plugin-commands-store/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @pnpm/plugin-commands-store
 
+## 8.1.9
+
+### Patch Changes
+
+- @pnpm/cli-utils@2.1.4
+- @pnpm/store-connection-manager@7.0.18
+
 ## 8.1.8
 
 ### Patch Changes